requirements.txt
acres==0.5.0
agate==1.9.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
alembic==1.14.0
amqp==5.3.1
aniso8601==9.0.1
annotated-types==0.7.0
ansicolors==1.1.8
antlr4-python3-runtime==4.9.3
anyio==4.7.0
asteroid-filterbanks==0.4.0
asttokens==3.0.0
asyncpg==0.30.0
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
daff==1.3.46
darkdetect==0.8.0
dataclasses-json==0.6.7
dbt-adapters==1.10.4
dbt-common==1.14.0
dbt-core==1.9.0
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.7.4
debugpy==1.8.14
decorator==5.2.1
deepdiff==7.0.1
Deprecated==1.2.15
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docutils==0.20.1
docx2txt==0.8
einops==0.8.1
entrypoints==0.4
etelemetry==0.3.1
eventlet==0.40.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
filetype==1.2.0
findspark==2.0.1
fitz==0.0.1.dev2
Flask==3.1.0
Flask-Alembic==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
flask-paginate==2024.4.12
flask-restx==1.3.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flower==2.0.1
fonttools==4.59.0
frontend==0.0.3
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.24.0
google-api-python-client==2.155.0
google-api-python-client-stubs==1.28.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.3
humanize==4.11.0
HyperPyYAML==1.2.2
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.8.1
kafka-python==2.0.2
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.8
kombu==5.4.2
langchain==0.3.11
langchain-core==0.3.24
langchain-text-splitters==0.3.2
langsmith==0.2.2
lazy_loader==0.4
leather==0.4.0
librosa==0.11.0
lightning==2.5.2
lightning-utilities==0.15.0
llama-cloud==0.1.6
llama-index==0.12.5
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.5
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.3.9
llama-index-multi-modal-llms-openai==0.4.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.1
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-postgres==0.4.0
llama-parse==0.5.17
llvmlite==0.41.1
looseversion==1.3.0
lxml==5.4.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mashumaro==3.14
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numba==0.58.1
numpy==1.26.2
omegaconf==2.3.0
openai==1.57.2
openai-whisper==20231117
optuna==4.4.0
ordered-set==4.1.0
orjson==3.10.12
packaging==23.2
pandas==2.2.3
papermill==2.6.0
parsedatetime==2.6
parso==0.8.4
pathlib==1.0.1
pathlib2==2.3.7.post1
pathspec==0.12.1
pgvector==0.3.6
pillow==11.0.0
platformdirs==4.3.8
pooch==1.8.2
primePy==1.3
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
prov==2.0.2
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
puremagic==1.29
py4j==0.10.9.7
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.21
pydantic==2.10.3
pydantic_core==2.27.1
pydot==4.0.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyodbc==5.2.0
pyparsing==3.2.0
pypdf==5.1.0
pypiwin32==223
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
python-engineio==4.10.1
python-slugify==8.0.4
python-socketio==5.11.4
pytimeparse==1.1.8
pytorch-lightning==2.5.2
pytorch-metric-learning==2.8.1
pytz==2024.2
pywin32==306
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.4.0
rdflib==7.1.4
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.1
scipy==1.15.3
semantic-version==2.10.0
semver==3.0.4
sentencepiece==0.2.0
setuptools-rust==1.8.1
shellingham==1.5.4
simple-websocket==1.1.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.0.4
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
speechbrain==1.0.3
SpeechRecognition==3.10.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.0
striprtf==0.0.26
subprocess.run==0.0.8
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
tensorboardX==2.6.4
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.8.0
tools==1.0.2
torch==2.7.1
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.1
torchmetrics==1.8.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
typer==0.16.0
types-httplib2==0.22.0.20240310
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.3
vine==5.1.0
virtualenv==20.25.0
Wave==0.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
